+++ packages/mwp-api-proxy-plugin/src/config.js
@@ -1,0 +1,2 @@
+export const API_PROXY_PLUGIN_NAME = 'mwp-api-proxy-plugin';
+export const API_ROUTE_PATH = '/mu_api';