+++ packages/mwp-core/src/util/globals.js
@@ -1,0 +1,6 @@
+import fetch from 'node-fetch';
+
+// hello polyfills
+global.URL = require('url').URL;
+global.URLSearchParams = require('url').URLSearchParams;
+global.fetch = fetch;